+++ gevent/_socket2.py
@@ -341,13 +341,16 @@ class socket(object):
             chunk = data_memory[data_sent:]
             if timeleft is None:
                 data_sent += self.send(chunk, flags)
+            elif timeleft <= 0:
+                # Check before sending to guarantee a check
+                # happens even if each chunk successfully sends its data
+                # (especially important for SSL sockets since they have large
+                # buffers)
+                raise timeout('timed out')
             else:
                 data_sent += self.send(chunk, flags, timeout=timeleft)
-                if data_sent >= len_data_memory:
-                    return
                 timeleft = end - time.time()
-                if timeleft <= 0:
-                    raise timeout('timed out')
+
         return timeleft
 
     def sendall(self, data, flags=0):