+++ gevent/_fileobjectcommon.py
@@ -1,4 +1,8 @@
-from gevent._socketcommon import EBADF
+
+try:
+    from errno import EBADF
+except ImportError:
+    EBADF = 9
 
 
 cancel_wait_ex = IOError(EBADF, 'File descriptor was closed in another greenlet')